re-did checkWebsite().
renamed it to isUsernameAvailable() since i read that i should minimize side effects in functions.

it made sense, so i did it. it now returns a boolean(unless there is a rate error)
i used try/except to deal with this when using the output of isUsernameAvailable()

also cleaned up comments